# api/docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: Ninaivalaigal Agentic API
  description: AI-powered memory management and context-aware API
  version: 1.0.0
  contact:
    name: Ninaivalaigal Team
    url: https://github.com/Arunosaur/ninaivalaigal
servers:
  - url: http://localhost:13370
    description: Development server (API container)
  - url: http://localhost:8000
    description: Development server (direct)
  - url: https://api.ninaivalaigal.com
    description: Production server
paths:
...
  /memory:
    get:
      summary: List all memories
      tags: [Memory]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: context
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of memories
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
                  total:
                    type: integer
    post:
      summary: Create a new memory
      tags: [Memory]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Memory content
                context:
                  type: string
                  description: Context name
                memory_type:
                  type: string
                  default: "user"
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Memory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
...
  /contexts:
    get:
      summary: List user contexts
      tags: [Contexts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of contexts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Context'
    post:
      summary: Create a new context
      tags: [Contexts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                scope:
                  type: string
                  enum: [personal, team, organization, global]
                  default: personal
      responses:
        '201':
          description: Context created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Memory:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        memory_type:
          type: string
        context:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        user_id:
          type: integer
        relevance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
    Context:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        scope:
          type: string
          enum: [personal, team, organization, global]
        is_active:
          type: boolean
        owner_id:
          type: integer
        created_at:
          type: string
          format: date-time
        memory_count:
          type: integer
...
security:
  - bearerAuth: []
...